+++ linux/auto_start_flutter_plugin_private.h
@@ -13,3 +13,4 @@ FlMethodResponse *get_device_manufacturer();
 FlMethodResponse *disable_battery_optimization();
 FlMethodResponse *is_battery_optimization_disabled();
 FlMethodResponse *open_app_info();
+FlMethodResponse *register_boot_callback();